More windows related shenanigans

+++ src-tauri/src/mediasession.rs
@@ -58,7 +58,7 @@ impl MediaSessionState {
         let hwnd_isize: isize = app
             .get_webview_window("main")
             .and_then(|w| w.hwnd().ok())
-            .map(|h| h.0)
+            .map(|h| h.0 as isize)
             .unwrap_or(0);
         #[cfg(not(target_os = "windows"))]
         let hwnd_isize: isize = 0;
